refactor(balance): meessage 매핑 전

+++ back/src/main/java/com/ssafy/db/repository/BalanceRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ssafy.db.entity.Balance;
 
 public interface BalanceRepository extends JpaRepository<Balance, Integer>{
-	Optional<Balance> findById(int id);
+	Optional<Balance> findById(long id);
 }
